requirements.txt
django>=4.0
djangorestframework>=3.14
django-crispy-forms>=2.0
crispy-bootstrap5>=2.0